requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
schedule>=1.2.0

# Optional (for paid features)
openai>=1.0.0

# HTML parsing
lxml>=4.9.0
html5lib>=1.1

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
